// management/frontend/src/api/schemas/workspaceMemberPublic.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */

export interface WorkspaceMemberPublic {
  id?: string;
  created_at?: string | null;
  updated_at?: string | null;
  deleted_at?: string | null;
  creator_id?: string | null;
  updater_id?: string | null;
  deleter_id?: string | null;
  user_id?: string | null;
  job_number?: string | null;
  employee_name?: string | null;
  sex?: string | null;
  mobile?: string | null;
  email?: string | null;
  telephone?: string | null;
  employee_status?: string | null;
  report_manager?: string | null;
  employee_type?: string | null;
  work_region?: string | null;
  company?: string | null;
  cooperation_mode?: string | null;
  work_serial_number?: string | null;
  height_cm?: number | null;
  position?: string | null;
  entry_date?: string | null;
  exit_date?: string | null;
  main_dept?: string | null;
  main_dept_id?: string | null;
  first_dept?: string | null;
  second_dept?: string | null;
  third_dept?: string | null;
  fourth_dept?: string | null;
  fifth_dept?: string | null;
  sixth_dept?: string | null;
  is_active?: boolean;
  account_id: string;
  workspace_id: string;
  creator_name?: string | null;
  updater_name?: string | null;
}