// management/frontend/src/api/zod/workspace-members/workspace-members.ts
export const workspaceMembersCreateWorkspaceMemberBodyUserIdOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodyJobNumberOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodyEmployeeNameOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodySexOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodyMobileOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodyEmailOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodyTelephoneOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodyEmployeeStatusOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodyReportManagerOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodyEmployeeTypeOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodyWorkRegionOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodyCompanyOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodyCooperationModeOneMax = 64;

export const workspaceMembersCreateWorkspaceMemberBodyWorkSerialNumberOneMax = 128;

export const workspaceMembersCreateWorkspaceMemberBodyHeightCmOneMin = 0;
export const workspaceMembersCreateWorkspaceMemberBodyHeightCmOneMax = 300;

export const workspaceMembersCreateWorkspaceMemberBodyPositionOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodyMainDeptOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodyMainDeptIdOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodyFirstDeptOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodySecondDeptOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodyThirdDeptOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodyFourthDeptOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodyFifthDeptOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodySixthDeptOneMax = 255;

export const workspaceMembersCreateWorkspaceMemberBodyIsActiveDefault = true;
export const workspaceMembersCreateWorkspaceMemberBodyUsernameMax = 150;

export const workspaceMembersCreateWorkspaceMemberBodyNewUserPasswordOneMin = 5;
export const workspaceMembersCreateWorkspaceMemberBodyNewUserPasswordOneMax = 128;



export const WorkspaceMembersCreateWorkspaceMemberBody = zod.object({
  "user_id": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodyUserIdOneMax),zod.null()]).optional(),
  "job_number": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodyJobNumberOneMax),zod.null()]).optional(),
  "employee_name": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodyEmployeeNameOneMax),zod.null()]).optional(),
  "sex": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodySexOneMax),zod.null()]).optional(),
  "mobile": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodyMobileOneMax),zod.null()]).optional(),
  "email": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodyEmailOneMax),zod.null()]).optional(),
  "telephone": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodyTelephoneOneMax),zod.null()]).optional(),
  "employee_status": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodyEmployeeStatusOneMax),zod.null()]).optional(),
  "report_manager": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodyReportManagerOneMax),zod.null()]).optional(),
  "employee_type": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodyEmployeeTypeOneMax),zod.null()]).optional(),
  "work_region": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodyWorkRegionOneMax),zod.null()]).optional(),
  "company": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodyCompanyOneMax),zod.null()]).optional(),
  "cooperation_mode": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodyCooperationModeOneMax),zod.null()]).optional(),
  "work_serial_number": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodyWorkSerialNumberOneMax),zod.null()]).optional(),
  "height_cm": zod.union([zod.number().min(workspaceMembersCreateWorkspaceMemberBodyHeightCmOneMin).max(workspaceMembersCreateWorkspaceMemberBodyHeightCmOneMax),zod.null()]).optional(),
  "position": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodyPositionOneMax),zod.null()]).optional(),
  "entry_date": zod.union([zod.iso.datetime({}),zod.null()]).optional(),
  "exit_date": zod.union([zod.iso.datetime({}),zod.null()]).optional(),
  "main_dept": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodyMainDeptOneMax),zod.null()]).optional(),
  "main_dept_id": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodyMainDeptIdOneMax),zod.null()]).optional(),
  "first_dept": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodyFirstDeptOneMax),zod.null()]).optional(),
  "second_dept": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodySecondDeptOneMax),zod.null()]).optional(),
  "third_dept": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodyThirdDeptOneMax),zod.null()]).optional(),
  "fourth_dept": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodyFourthDeptOneMax),zod.null()]).optional(),
  "fifth_dept": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodyFifthDeptOneMax),zod.null()]).optional(),
  "sixth_dept": zod.union([zod.string().max(workspaceMembersCreateWorkspaceMemberBodySixthDeptOneMax),zod.null()]).optional(),
  "is_active": zod.boolean().default(workspaceMembersCreateWorkspaceMemberBodyIsActiveDefault),
  "username": zod.string().min(1).max(workspaceMembersCreateWorkspaceMemberBodyUsernameMax),
  "new_user_password": zod.union([zod.string().min(workspaceMembersCreateWorkspaceMemberBodyNewUserPasswordOneMin).max(workspaceMembersCreateWorkspaceMemberBodyNewUserPasswordOneMax),zod.null()]).optional()
})
// ...
export const workspaceMembersUpdateWorkspaceMemberBodyUserIdOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodyJobNumberOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodyEmployeeNameOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodySexOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodyMobileOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodyEmailOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodyTelephoneOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodyEmployeeStatusOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodyReportManagerOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodyEmployeeTypeOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodyWorkRegionOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodyCompanyOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodyCooperationModeOneMax = 64;

export const workspaceMembersUpdateWorkspaceMemberBodyWorkSerialNumberOneMax = 128;

export const workspaceMembersUpdateWorkspaceMemberBodyHeightCmOneMin = 0;
export const workspaceMembersUpdateWorkspaceMemberBodyHeightCmOneMax = 300;

export const workspaceMembersUpdateWorkspaceMemberBodyPositionOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodyMainDeptOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodyMainDeptIdOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodyFirstDeptOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodySecondDeptOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodyThirdDeptOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodyFourthDeptOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodyFifthDeptOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodySixthDeptOneMax = 255;

export const workspaceMembersUpdateWorkspaceMemberBodyIsActiveDefault = true;

export const WorkspaceMembersUpdateWorkspaceMemberBody = zod.object({
  "user_id": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodyUserIdOneMax),zod.null()]).optional(),
  "job_number": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodyJobNumberOneMax),zod.null()]).optional(),
  "employee_name": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodyEmployeeNameOneMax),zod.null()]).optional(),
  "sex": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodySexOneMax),zod.null()]).optional(),
  "mobile": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodyMobileOneMax),zod.null()]).optional(),
  "email": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodyEmailOneMax),zod.null()]).optional(),
  "telephone": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodyTelephoneOneMax),zod.null()]).optional(),
  "employee_status": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodyEmployeeStatusOneMax),zod.null()]).optional(),
  "report_manager": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodyReportManagerOneMax),zod.null()]).optional(),
  "employee_type": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodyEmployeeTypeOneMax),zod.null()]).optional(),
  "work_region": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodyWorkRegionOneMax),zod.null()]).optional(),
  "company": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodyCompanyOneMax),zod.null()]).optional(),
  "cooperation_mode": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodyCooperationModeOneMax),zod.null()]).optional(),
  "work_serial_number": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodyWorkSerialNumberOneMax),zod.null()]).optional(),
  "height_cm": zod.union([zod.number().min(workspaceMembersUpdateWorkspaceMemberBodyHeightCmOneMin).max(workspaceMembersUpdateWorkspaceMemberBodyHeightCmOneMax),zod.null()]).optional(),
  "position": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodyPositionOneMax),zod.null()]).optional(),
  "entry_date": zod.union([zod.iso.datetime({}),zod.null()]).optional(),
  "exit_date": zod.union([zod.iso.datetime({}),zod.null()]).optional(),
  "main_dept": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodyMainDeptOneMax),zod.null()]).optional(),
  "main_dept_id": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodyMainDeptIdOneMax),zod.null()]).optional(),
  "first_dept": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodyFirstDeptOneMax),zod.null()]).optional(),
  "second_dept": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodySecondDeptOneMax),zod.null()]).optional(),
  "third_dept": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodyThirdDeptOneMax),zod.null()]).optional(),
  "fourth_dept": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodyFourthDeptOneMax),zod.null()]).optional(),
  "fifth_dept": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodyFifthDeptOneMax),zod.null()]).optional(),
  "sixth_dept": zod.union([zod.string().max(workspaceMembersUpdateWorkspaceMemberBodySixthDeptOneMax),zod.null()]).optional(),
  "is_active": zod.boolean().default(workspaceMembersUpdateWorkspaceMemberBodyIsActiveDefault)
})